requirements.txt
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mistletoe==0.9.0
numpy==1.23.2
openai==0.23.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
pandas-stubs==1.4.3.220829
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
tqdm==4.64.0
transformers==4.22.1
types-pytz==2022.2.1.0
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
